requirements.txt
# AI agent layer
openai>=1.50.0

# Slack bot
slack-bolt>=1.21.0
slack-sdk>=3.33.0

# HTTP / API clients
requests>=2.32.0
httpx>=0.27.0

# News & web research
newsapi-python>=0.2.7
tavily-python>=0.5.0

# Environment
python-dotenv>=1.0.0

# Data handling
pandas>=2.2.0
